ATR-556: refactored DTE events observing logic into separate private nested class to group realted logic together and improve main view model readability
